update parameters to proven defaults
With these parameters, 1225 generations gave:

Total: 8159
0: 94.29%
1: 97.89%
2: 84.21%
3: 87.43%
4: 89.41%
5: 00.00%
6: 96.56%
7: 91.44%
8: 87.78%
9: 76.71%

That is 1.31 epochs. Not as good as SGD, but getting somewhere. It looks like a majority of the problem is that 5 is never predicted. The bias for that neuron may have gotten a little bit too low to recover, but I can't know because I lost the final network.