resist statistical issues with weight decay
Weight decay provides a nice way to prevent the parameters in the network from blowing up. Basically, if we use a certain amount of weight decay, we can target a specific parameter standard deviation. In this case, I've set everything up to aim for a stddev of 1 for all weights and biases. Then, I restructured the network so that such a setup would actually work well.

TODO: restructure the convolutional net to benefit from stddev=1.