added regular SGD comparison
In about 11 seconds, SGD gets the network up to 95% accuracy. Compare this to the few hours that evolution spent getting a network up to 84% accuracy. That's the price of having to deal with an entire population :\

One thing to note is that the variances of the trained network's parameters are all < 0.09. This indicates that the parameter explosion from evolution might be the cause of the learning stagnation.