automatic decay scheduling
This took a lot of work to implement, but the code itself is only a few lines.

The code for generating data is here: https://gist.github.com/unixpickle/20bbb4721b81bf75aada73276eb9d417

The function of best fit was found using octave. The solution is a least squares solution in the log domain, acting on log(noise), log(decay), and 1.